pi/requirements.txt
paho-mqtt==2.1.0
qrcode[pil]==7.4.2
Pillow==10.3.0
brother_ql==0.9.4
RPi.GPIO==0.7.1
spidev==3.6
